Refactor into running_tracker package
