+++ main.py
@@ -1,4 +1,4 @@
-from gui import RunningTrackerApp
+from running_tracker import RunningTrackerApp
 
 if __name__ == "__main__":
     app = RunningTrackerApp()
