+++ running_tracker/__init__.py
@@ -1,0 +1,3 @@
+from .gui import RunningTrackerApp
+
+__all__ = ["RunningTrackerApp"]
